chore(no-commented-out-tests): migrate to TS
